调整汪汪寻宝三连概率
1. 重新设计汪汪寻宝单格图标权重，提高奖励图标的单格出现率。

2. 将三连奖励概率按单格权重的三次方导出，并支持小数万分比展示。

3. 更新寻宝概率和抽奖边界测试，覆盖新的权重语义。

+++ xechat-commons/src/main/java/cn/xeblog/commons/entity/pet/PetTreasureHuntProbabilityDTO.java
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 import java.io.Serializable;
 
 /**
- * 汪汪寻宝概率展示项，probabilityBp 为万分比。
+ * 汪汪寻宝概率展示项，probabilityBp 为万分比，可包含小数。
  */
 @Data
 @NoArgsConstructor
@@ -20,7 +20,7 @@ public class PetTreasureHuntProbabilityDTO implements Serializable {
 
     private String label;
 
-    private int probabilityBp;
+    private double probabilityBp;
 
     private Integer value;
